I fixed the memory leak and got it working, finally!
It adds the signature to all the photos and does so in <1 sec/image (The debugging println statements are the main limiting factor)
It turns out the memory leak was somehow caused by incompatibilities when converting APImage objects to BufferedImage ones, which I was able to circumvent by replacing all APImage objects with BufferedImage ones, which was easy, as I'd done so with a few variables earlier which had caused other problems.
Now all that's left is for some stylistic and formatting changes to be made to make the signature look better and be more visible.